submissions/2026-04-29__001_synthetic_mine_throughput__opencode__gemini-3-1-pro-preview__customtools-superpowers/requirements.txt
simpy>=4.1
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
networkx>=3.0
pyyaml>=6.0
pytest>=7.0
jsonschema>=4.0

